\name{listThesisExamples}
\alias{listThesisExamples}
\title{
  List available examples
}
\description{
  Print out a list of the available examples. The name of each example
  can be given to \code{\link{thesisExample}} to be started, and then
  viewed within a web browser.
}
\usage{
listThesisExamples()
}
\value{
  None.
}
\author{
  Simon Potter
}
\seealso{
  \code{\link{thesisExample}}
}
\examples{\dontrun{
listThesisExamples()
}}
